requirements.txt
Flask==3.0.0
gunicorn==20.1.0
google-cloud-datastore==2.15.1
google-auth==2.17.3
requests==2.28.2
flask-login
openai